Updated tests for circleci build
